refactor: extract staking input logic to hook
