+++ src/components/common/ColonyActions/ActionDetailsPage/DefaultMotion/MotionPhaseWidget/StakingWidget/StakingInput/index.ts
@@ -2,6 +2,7 @@ export {
   default,
   SLIDER_AMOUNT_KEY,
   getStakingTransformFn,
+  StakingWidgetValues,
 } from './StakingInput';
 export { default as StakingControls } from './StakingControls';
 export { default as StakingSliderLabel } from './StakingSliderLabel';
