+++ src/components/shared/Fields/Form/index.ts
@@ -1,4 +1,4 @@
 export { default } from './Form';
 export { default as HookForm } from './HookForm';
 export { default as ActionForm } from './ActionForm';
-export { default as ActionHookForm } from './ActionHookForm';
+export { default as ActionHookForm, OnSuccess } from './ActionHookForm';
