+++ src/components/shared/Fields/index.ts
@@ -1,6 +1,12 @@
 export { default as AmountTokens } from './AmountTokens';
 export { default as Checkbox, HookFormCheckbox } from './Checkbox';
-export { default as Form, ActionForm, HookForm, ActionHookForm } from './Form';
+export {
+  default as Form,
+  ActionForm,
+  HookForm,
+  ActionHookForm,
+  OnSuccess,
+} from './Form';
 export { default as FormStatus } from './FormStatus';
 export { default as FieldSet } from './FieldSet';
 export { default as Input, InputComponentAppearance } from './Input';
